Split SQS logic into own class
